# backend/docs/openapi.yaml
openapi: 3.0.3
info:
  title: ReviewApp API
  description: |
    **「使えば使うほど、あなたらしくなるAIアシスタント」**
    
    RAGベースのAIコードレビューツール。
    あなたのコーディング哲学を学習し、一貫性のある「あなたらしい」レビューを提供します。
    
    ## 主要機能
    - 📚 **ナレッジ管理**: コーディングルール・哲学を蓄積
    - 🔍 **RAG検索**: Phase 1（キーワード） → Phase 2（ベクトル検索）
    - 💬 **コードレビュー**: 過去のナレッジを参照した一貫性のあるレビュー
    - 📊 **レビュー履歴**: フィードバックで精度向上
    
  version: 0.1.0
  contact:
    name: ReviewApp Support
    email: support@reviewapp.example.com
  license:
    name: MIT
servers:
  - url: http://localhost:8080
    description: ローカル開発環境
  - url: https://api.reviewapp.example.com
    description: 本番環境
...
security:
  - BearerAuth: []
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT認証トークン

  # =====================================================
  # 共通スキーマ
  # =====================================================
  schemas:
...
    Error:
      type: object
      required:
        - error
        - message
      properties:
        error:
          type: string
          example: "invalid_request"
        message:
          type: string
          example: "リクエストが不正です"
        details:
          type: object
          additionalProperties: true

    # --- ページネーション ---
    Pagination:
      type: object
      properties:
        page:
          type: integer
          example: 1
        per_page:
          type: integer
          example: 20
        total:
          type: integer
          example: 100
        total_pages:
          type: integer
          example: 5

    # --- User ---
...
    Knowledge:
      type: object
      required:
        - id
        - user_id
        - title
        - content
        - category
        - priority
        - source_type
        - is_active
        - created_at
      properties:
        id:
          type: string
          format: uuid
        user_id:
          type: string
          format: uuid
        title:
          type: string
          maxLength: 200
          example: "エラーハンドリングの原則"
        content:
          type: string
          example: "エラーは必ずログに出力し、ユーザー向けメッセージと開発者向け詳細を分ける"
        category:
          type: string
          enum:
            - error_handling
            - testing
            - performance
            - security
            - clean_code
            - architecture
            - other
          example: "error_handling"
        priority:
          type: integer
          minimum: 1
          maximum: 5
          example: 5
          description: "重要度（1=低、5=高）"
        tags:
          type: array
          items:
            $ref: '#/components/schemas/Tag'
        source_type:
          type: string
          enum:
            - review
            - conversation
            - manual
          example: "manual"
        source_id:
          type: string
          format: uuid
          nullable: true
        is_active:
          type: boolean
          example: true
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time

    # --- Knowledge Input ---
    KnowledgeInput:
      type: object
      required:
        - title
        - content
        - category
      properties:
        title:
          type: string
          maxLength: 200
          example: "エラーハンドリングの原則"
        content:
          type: string
          example: "エラーは必ずログに出力する"
        category:
          type: string
          example: "error_handling"
        priority:
          type: integer
          minimum: 1
          maximum: 5
          default: 3
        tags:
          type: array
          items:
            type: string
          example: ["go", "error"]

    # --- Review ---
...
    Tag:
      type: object
      required:
        - id
        - name
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
          example: "go"
        color:
          type: string
          pattern: '^#[0-9A-Fa-f]{6}$'
          example: "#3178C6"
        created_at:
          type: string
          format: date-time

# =====================================================
# Paths（エンドポイント）
# =====================================================
paths:
...
  /api/v1/knowledge:
    get:
      tags:
        - Knowledge
      summary: ナレッジ一覧を取得
      parameters:
        - name: category
          in: query
          description: カテゴリでフィルタ
          schema:
            type: string
            example: "error_handling"
        - name: tag
          in: query
          description: タグでフィルタ
          schema:
            type: string
            example: "go"
        - name: priority
          in: query
          description: 優先度でフィルタ
          schema:
            type: integer
            minimum: 1
            maximum: 5
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: per_page
          in: query
          schema:
            type: integer
            default: 20
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Knowledge'
                  pagination:
                    $ref: '#/components/schemas/Pagination'

    post:
      tags:
        - Knowledge
      summary: ナレッジを作成
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/KnowledgeInput'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Knowledge'
        '400':
          description: バリデーションエラー
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...